+++ app.js
@@ -185,7 +185,7 @@ client.on('message', msg => {
                 new MarioMaker(msg);
             } else if (input.startsWith("ssbu")) {
                 new Arena(msg);
-            } else if (input.startsWith("acnh") || input.startsWith("dodo")) {
+            } else if (input.startsWith("acnh") || input.startsWith("dodo") || input.startsWith("anch")) {
                 new DodoCode(msg);
             } else if (input.startsWith("donate")) {
                 msg.channel.send({